Add button to open context menu (#728)
* Add button to open context menu

In some contexts, it's hard or impossible to open the context menu of the emulator, which is required to access important secondary options such as quick states, screen recording, and software information. Primarily, it's apparently impossible to open the context menu on a mobile web browser, since a long press on the screen is, for good reason, not registered like a right mouse click is on desktop. This commit adds what I think is the ideal solution to the issue: a button on the menu bar/drawer that opens up the context menu, without cluttering it up with multiple options, just a single new one.

* Update en locale (en.json, Translate.html)

* Fix contextMenuButton dimensions

* Set contextMenu to be centered when open via button