Add support for Arcade BIOS files
The default logic decompresses the BIOS file if it's a zip. On arcade systems, it is quite normal to have BIOS files _being_ an actual zip file. Decompressing it makes the game loading fail. This approach stores the zip file before decompressing, if it's an arcade or mame system, and then proceeds to decompressing - in case it's a zip with multiple zip files for some more complex systems with multiple drivers.
The proper solution would be for us to allow for multiple BIOS attachments, but this compromise would work.